fix(pagination): Sets real paginated results for cards list

+++ src/routes/responses/mod.rs
@@ -17,7 +17,7 @@ pub struct ListSuccessMeta {
 
 #[derive(serde::Serialize, serde::Deserialize)]
 pub struct PaginatedListSuccessMeta {
-    pub total: usize,
+    pub total: i64,
     pub success: bool,
     pub next_page: Option<String>,
     pub previous_page: Option<String>,
